add categorical anyvalue
