+++ polars/polars-core/src/chunked_array/ops/any_value.rs
@@ -1,4 +1,69 @@
+#[cfg(feature = "dtype-categorical")]
+use crate::chunked_array::categorical::RevMapping;
+#[cfg(not(feature = "dtype-categorical"))]
+use crate::chunked_array::RevMapping;
 use crate::prelude::*;
+use std::convert::TryFrom;
+
+#[inline]
+#[allow(unused_variables)]
+unsafe fn arr_to_any_value<'a>(
+    arr: &'a dyn Array,
+    idx: usize,
+    categorical_map: &'a Option<Arc<RevMapping>>,
+    dtype: &DataType,
+) -> AnyValue<'a> {
+    if arr.is_null(idx) {
+        return AnyValue::Null;
+    }
+
+    macro_rules! downcast_and_pack {
+        ($casttype:ident, $variant:ident) => {{
+            let arr = &*(arr as *const dyn Array as *const $casttype);
+            let v = arr.value(idx);
+            AnyValue::$variant(v)
+        }};
+    }
+    macro_rules! downcast {
+        ($casttype:ident) => {{
+            let arr = &*(arr as *const dyn Array as *const $casttype);
+            arr.value_unchecked(idx)
+        }};
+    }
+    // TODO: insert types
+    match dtype {
+        DataType::Utf8 => downcast_and_pack!(LargeStringArray, Utf8),
+        DataType::Boolean => downcast_and_pack!(BooleanArray, Boolean),
+        DataType::UInt8 => downcast_and_pack!(UInt8Array, UInt8),
+        DataType::UInt16 => downcast_and_pack!(UInt16Array, UInt16),
+        DataType::UInt32 => downcast_and_pack!(UInt32Array, UInt32),
+        DataType::UInt64 => downcast_and_pack!(UInt64Array, UInt64),
+        DataType::Int8 => downcast_and_pack!(Int8Array, Int8),
+        DataType::Int16 => downcast_and_pack!(Int16Array, Int16),
+        DataType::Int32 => downcast_and_pack!(Int32Array, Int32),
+        DataType::Int64 => downcast_and_pack!(Int64Array, Int64),
+        DataType::Float32 => downcast_and_pack!(Float32Array, Float32),
+        DataType::Float64 => downcast_and_pack!(Float64Array, Float64),
+        #[cfg(feature = "dtype-date32")]
+        DataType::Date32 => downcast_and_pack!(Int32Array, Date32),
+        #[cfg(feature = "dtype-date64")]
+        DataType::Date64 => downcast_and_pack!(Int64Array, Date64),
+        DataType::List(_) => {
+            let v: ArrayRef = downcast!(LargeListArray).into();
+            let s = Series::try_from(("", v));
+            AnyValue::List(s.unwrap())
+        }
+        #[cfg(feature = "dtype-categorical")]
+        DataType::Categorical => {
+            let idx = downcast!(UInt32Array);
+            let rev_map = &**categorical_map.as_ref().unwrap();
+            AnyValue::Categorical(idx, rev_map)
+        }
+        #[cfg(feature = "object")]
+        DataType::Object(_) => panic!("should not be here"),
+        _ => unimplemented!(),
+    }
+}
 
 macro_rules! get_any_value_unchecked {
     ($self:ident, $index:expr) => {{
@@ -6,7 +71,7 @@ macro_rules! get_any_value_unchecked {
         debug_assert!(chunk_idx < $self.chunks.len());
         let arr = &**$self.chunks.get_unchecked(chunk_idx);
         debug_assert!(idx < arr.len());
-        $self.arr_to_any_value(arr, idx)
+        arr_to_any_value(arr, idx, &$self.categorical_map, $self.dtype())
     }};
 }
 
@@ -17,7 +82,7 @@ macro_rules! get_any_value {
         assert!(idx < arr.len());
         // SAFETY
         // bounds are checked
-        unsafe { $self.arr_to_any_value(arr, idx) }
+        unsafe { arr_to_any_value(arr, idx, &$self.categorical_map, $self.dtype()) }
     }};
 }
